+++ noodles-util/examples/util_alignment_rewrite.rs
@@ -4,13 +4,12 @@
 
 use std::{
     env,
-    ffi::{OsStr, OsString},
     fs::File,
-    io::{self, BufReader, BufWriter},
-    path::{Path, PathBuf},
+    io::{self, BufWriter},
+    path::Path,
 };
 
-use noodles_fasta::{self as fasta, fai, repository::adapters::IndexedReader};
+use noodles_fasta::{self as fasta, repository::adapters::IndexedReader};
 use noodles_util::alignment::{self, Format};
 
 fn detect_format_from_extension<P>(path: P) -> Option<Format>
@@ -25,34 +24,6 @@ where
     }
 }
 
-fn push_ext<S>(path: PathBuf, ext: S) -> PathBuf
-where
-    S: AsRef<OsStr>,
-{
-    let mut s = OsString::from(path);
-    s.push(".");
-    s.push(ext);
-    PathBuf::from(s)
-}
-
-fn create_reference_sequence_repository<P>(src: P) -> io::Result<fasta::Repository>
-where
-    P: AsRef<Path>,
-{
-    let src = src.as_ref();
-
-    let reader = File::open(src)
-        .map(BufReader::new)
-        .map(fasta::Reader::new)?;
-
-    let fai_src = push_ext(src.to_path_buf(), "fai");
-    let index = fai::read(fai_src)?;
-
-    let adapter = IndexedReader::new(reader, index);
-
-    Ok(fasta::Repository::new(adapter))
-}
-
 fn main() -> io::Result<()> {
     let mut args = env::args().skip(1);
 
@@ -61,8 +32,9 @@ fn main() -> io::Result<()> {
     let fasta_src = args.next();
 
     let repository = fasta_src
-        .map(create_reference_sequence_repository)
+        .map(|src| IndexedReader::builder().open(src))
         .transpose()?
+        .map(fasta::Repository::new)
         .unwrap_or_default();
 
     let mut reader = File::open(src).and_then(|f| {